pyproject.toml
[project]
name = "grogue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
]
